+++ rector.php
@@ -3,6 +3,8 @@
 use Rector\Caching\ValueObject\Storage\FileCacheStorage;
 use Rector\Config\RectorConfig;
 use Rector\Php81\Rector\Array_\FirstClassCallableRector;
+use Rector\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector;
+use Rector\Php84\Rector\MethodCall\NewMethodCallWithoutParenthesesRector;
 use RectorLaravel\Rector\Class_\UnifyModelDatesWithCastsRector;
 use RectorLaravel\Rector\FuncCall\RemoveDumpDataDeadCodeRector;
 use RectorLaravel\Rector\MethodCall\EloquentWhereRelationTypeHintingParameterRector;
@@ -41,6 +43,16 @@ return RectorConfig::configure()
         // transforme :     array_map('intval',
         // en :             array_map(intval(...),
         FirstClassCallableRector::class,
+
+        // Cet attribut natif PHP n'est pas très utile ;
+        // mieux vaux se baser sur de l'analyse statique
+        // En plus, lors de la rédaction de ce message,
+        // la règle fonctionne mal, elle est buguée...
+        AddOverrideAttributeToOverriddenMethodsRector::class,
+
+        // Tant que PHP-CS-Fixer ne supporte pas PHP 8.4,
+        // pour que PINT fonctionne correctement
+        NewMethodCallWithoutParenthesesRector::class,
     ])
     ->withRules([
         EloquentWhereRelationTypeHintingParameterRector::class,